Create file '.gitignore' and remove redundant files and folders.
There's an IDEA plugin called '.ignore' and can be used to create most common pattens of ignore files.

Addition operations:
git rm --cached Pos.iml
git rm --cached Pos.ipr
git rm --cached Pos.iws
git rm -r --cached .gradle
git rm -r --cached build
git rm -r --cached .idea

The iml(IDEA module), ipr(IDEA project) and iws(IDEA workspace) files could be produced in 'gradle idea' as we're using Gradle for project settings and should not be checked in VCS.
The .gradle directory is for cache purpose, its contents could be auto achieved in gradle tasks and shouldn't be checked in VCS.
The .idea directory is for cache of operations of Intellij IDEA(eg. which line in which file you're editting at last) and is of no use when out of the IDE, shouldn't be checked in VCS.
The build directory is the default target directory of the build task in gradle, all its content could be generated in 'gradle build' command and shouldn't be checked in VCS.